Move log package to the right location, fix shouldLog typo
