+++ src/test/java/net/fabricmc/test/EntrypointTest.java
@@ -16,8 +16,8 @@
 
 package net.fabricmc.test;
 
-import net.fabricmc.loader.util.log.Log;
-import net.fabricmc.loader.util.log.LogCategory;
+import net.fabricmc.loader.impl.util.log.Log;
+import net.fabricmc.loader.impl.util.log.LogCategory;
 
 public final class EntrypointTest {
 	public static final CustomEntry FIELD_ENTRY = EntrypointTest::fieldEntry;
